Refactor the initial filter/sort/select settings based on url to be managed at the PaletteContainer level rather than the select level.
Because on mobile view the select fields are hidden by default the
filters (being applied by url only on mount) would be unapplied until
the user clicked the dropdown. This forced a fix where the menu
defaulted to open. This provided a poor ux and has now been fixed by
lifting the initial filter/sort/search url param checking logic to the
PaletteContainer's mount LC method.
